Split code for all and approved dApps tabs
