+++ app/src/main/java/com/babylon/wallet/android/presentation/dappdir/all/AllDAppsViewModel.kt
@@ -1,0 +1,64 @@
+package com.babylon.wallet.android.presentation.dappdir.all
+
+import androidx.lifecycle.viewModelScope
+import com.babylon.wallet.android.presentation.common.StateViewModel
+import com.babylon.wallet.android.presentation.dappdir.common.delegates.DAppListDelegate
+import com.babylon.wallet.android.presentation.dappdir.common.models.DAppListItem.DAppWithDetails
+import com.babylon.wallet.android.presentation.dappdir.common.models.DAppListState
+import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.combine
+import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.flow.update
+import javax.inject.Inject
+import kotlin.collections.map
+import kotlin.collections.orEmpty
+
+@HiltViewModel
+class AllDAppsViewModel @Inject constructor(
+    private val dAppListDelegate: DAppListDelegate
+) : StateViewModel<DAppListState>(), DAppListDelegate.ViewActions by dAppListDelegate {
+
+    private val dAppsWithDetails: Flow<List<DAppWithDetails>> = combine(
+        dAppListDelegate.directoryState.map { directory ->
+            directory?.copy(
+                // Shuffle order of highlighted dApps
+                highlighted = directory.highlighted?.shuffled(),
+                // Shuffle order of other dApps
+                others = directory.others?.shuffled()
+            )
+        },
+        dAppListDelegate.dAppDataState
+    ) { directory, dAppData ->
+        directory?.highlighted.orEmpty().map {
+            DAppWithDetails(
+                dAppDefinitionAddress = it.dAppDefinitionAddress,
+                details = dAppData.getOrDefault(it.dAppDefinitionAddress, DAppWithDetails.Details.Fetching)
+            )
+        } + directory?.others.orEmpty().map {
+            DAppWithDetails(
+                dAppDefinitionAddress = it.dAppDefinitionAddress,
+                details = dAppData.getOrDefault(it.dAppDefinitionAddress, DAppWithDetails.Details.Fetching)
+            )
+        }
+    }
+
+    init {
+        dAppListDelegate.initialize(
+            scope = viewModelScope,
+            state = _state,
+            dAppsWithDetailsState = dAppsWithDetails
+        )
+    }
+
+    override fun initialState(): DAppListState = DAppListState(isLoading = true)
+
+    fun onRefresh() {
+        _state.update { it.copy(isRefreshing = true) }
+        dAppListDelegate.loadDAppDirectory()
+    }
+
+    fun onMessageShown() {
+        _state.update { it.copy(uiMessage = null) }
+    }
+}
