+++ app/src/main/java/com/babylon/wallet/android/presentation/dappdir/common/models/DAppFilters.kt
@@ -1,0 +1,10 @@
+package com.babylon.wallet.android.presentation.dappdir.common.models
+
+data class DAppFilters(
+    val searchTerm: String = "",
+    val selectedTags: Set<String> = emptySet(),
+    val availableTags: Set<String> = emptySet()
+) {
+
+    fun isTagSelected(tag: String) = tag in selectedTags
+}
